+++ src/components/ThemePicker/index.ts
@@ -1,0 +1,3 @@
+import ThemePicker from './ThemePicker';
+
+export { ThemePicker };
