--- src/pages/meditation/components/MeditationThemePicker/index.ts
@@ -1,3 +1,0 @@
-import MeditationThemePicker from './MeditationThemePicker';
-
-export { MeditationThemePicker };
